first Draft 2 endings
copied a passage of the erverything api script and managed to make the 2nd ending work
add the screens and animations for the 2 endings

the audio gets cut off though and the anim clip only shows up real short before going back to the normal one.
also: if time -> make it so that if the glitch effect is in use, the rest of the animations and clips can't overwrite it
still need to add colours, glitch sounds etc.